+++ geo/src/algorithm/line_measures/metric_spaces/haversine.rs
@@ -4,44 +4,47 @@ use super::super::{Bearing, Destination, Distance, InterpolatePoint};
 use crate::utils::normalize_longitude;
 use crate::{CoordFloat, Point, MEAN_EARTH_RADIUS};
 
-/// A spherical model of the earth using the [haversine formula].
+/// Use the [`Haversine`] constant (an instance of `HaversineMeasure`) rather than building your own
+/// customized [`HaversineMeasure`] for standard spherical Earth measurements.
 ///
-/// Distances are considered [great circle] lengths and are measured in meters.
+/// `HaversineMeasure` measures distance on a sphere using the [haversine formula].
+/// Distances are considered [great circle] lengths and given in units that match those of
+/// the `radius` passed to [`HaversineMeasure::new`] (typically meters).
 ///
 /// You may specify a custom radius for the Earth (or other sphere), but for normal spherical
-/// measurements of the Earth, you probably want to just use [`HAVERSINE`] which uses the
+/// measurements of the Earth, you should just use [`Haversine`] which uses the standard
 /// earth radius of 6371.0088 km (6_371_008.7714 m), based on the recommendation of the IUGG.
 ///
 /// # Examples
 ///
 /// ```
 /// # use approx::assert_relative_eq;
-/// use geo::{wkt, Haversine, HAVERSINE, Distance};
+/// use geo::{wkt, HaversineMeasure, Haversine, Distance};
 ///
 /// let start = wkt!(POINT(23.319941 42.698334)); // Sofia: Longitude, Latitude
 /// let finish = wkt!(POINT(24.742168 42.136097)); // Plovdiv: Longitude, Latitude
 ///
-/// // Typically, you can just use `HAVERSINE` for measuring on the Earth's surface
+/// // Typically, you can just use `Haversine` for measuring on the Earth's surface.
 /// assert_relative_eq!(
 ///     132433.09929460194,
-///     HAVERSINE.distance(start, finish)
+///     Haversine.distance(start, finish)
 /// );
 ///
-/// // `HAVERSINE` has a radius equal to the mean radius of the GRS80 ellipsoid.
+/// // The default `Haversine` const has a radius equal to the mean radius of the GRS80 ellipsoid (6371.0088 km).
 /// assert_relative_eq!(
-///     HAVERSINE.distance(start, finish),
-///     Haversine::GRS80_MEAN_RADIUS.distance(start, finish)
+///     Haversine.radius(),
+///     HaversineMeasure::GRS80_MEAN_RADIUS.radius()
 /// );
 ///
 /// // You may choose to use one of the other well known estimations of the Earth's radius,
 /// // which may result in *slightly* different results.
 /// assert_relative_eq!(
 ///     132433.06564071847,
-///     Haversine::GRS80_EQUAL_AREA.distance(start, finish)
+///     HaversineMeasure::GRS80_EQUAL_AREA.distance(start, finish)
 /// );
 ///
 /// // Or you can specify whatever radius you want to get some "out of this world" results.
-/// let mars_sphere = Haversine::new(3_389_500.0); // 👽 Mars radius in meters
+/// let mars_sphere = HaversineMeasure::new(3_389_500.0); // 👽 Mars radius in meters
 /// assert_relative_eq!(
 ///     70456.97222377927,
 ///     mars_sphere.distance(start, finish)
@@ -55,22 +58,52 @@ use crate::{CoordFloat, Point, MEAN_EARTH_RADIUS};
 /// - <https://link.springer.com/article/10.1007%2Fs001900050278>
 /// - <https://sci-hub.se/https://doi.org/10.1007/s001900050278>
 ///
-/// [haversine formula]: https://en.wikipedia.org/wiki/Haversine_formula//
+/// [haversine formula]: https://en.wikipedia.org/wiki/Haversine_formula
 /// [great circle]: https://en.wikipedia.org/wiki/Great_circle
-pub struct Haversine {
+pub struct HaversineMeasure {
     radius: f64,
 }
 
-impl Default for Haversine {
+impl Default for HaversineMeasure {
     fn default() -> Self {
-        Haversine::GRS80_MEAN_RADIUS
+        HaversineMeasure::GRS80_MEAN_RADIUS
     }
 }
 
-/// A spherical model of the earth using the [haversine formula], and the GRS80 Mean Earth Radius of 6371008.8 meters.
-pub const HAVERSINE: Haversine = Haversine::GRS80_MEAN_RADIUS;
+/// `Haversine` measures distance on a sphere using the [haversine formula]. Distances are
+/// considered [great circle] lengths and given in meters.
+///
+/// See [`HaversineMeasure`](HaversineMeasure#trait-implementations) for all the trait methods available to [`Haversine`].
+///
+// / # Examples
+///
+/// ```
+/// # use approx::assert_relative_eq;
+/// use geo::{wkt, Haversine, Distance};
+///
+/// let start = wkt!(POINT(23.319941 42.698334)); // Sofia: Longitude, Latitude
+/// let finish = wkt!(POINT(24.742168 42.136097)); // Plovdiv: Longitude, Latitude
+///
+/// // Typically, you can just use `Haversine` for measuring on the Earth's surface.
+/// assert_relative_eq!(
+///     132433.09929460194,
+///     Haversine.distance(start, finish)
+/// );
+/// ```
+///
+/// # References
+///
+/// Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278
+/// "Derived Geometric Constants: **R1: mean radius**" (p131)
+/// - <https://link.springer.com/article/10.1007%2Fs001900050278>
+/// - <https://sci-hub.se/https://doi.org/10.1007/s001900050278>
+///
+/// [haversine formula]: https://en.wikipedia.org/wiki/Haversine_formula
+/// [great circle]: https://en.wikipedia.org/wiki/Great_circle
+#[allow(non_upper_case_globals)]
+pub const Haversine: HaversineMeasure = HaversineMeasure::GRS80_MEAN_RADIUS;
 
-impl Haversine {
+impl HaversineMeasure {
     /// ## Parameters
     /// - radius: The radius of the sphere, typically in meters.
     pub const fn new(radius: f64) -> Self {
@@ -121,13 +154,7 @@ impl Haversine {
     };
 }
 
-impl Default for Haversine {
-    fn default() -> Self {
-        Self::GRS80_MEAN_RADIUS
-    }
-}
-
-impl<F: CoordFloat + FromPrimitive> Bearing<F> for Haversine {
+impl<F: CoordFloat + FromPrimitive> Bearing<F> for HaversineMeasure {
     /// Returns the bearing from `origin` to `destination` in degrees along a [great circle].
     ///
     /// # Units
@@ -139,12 +166,12 @@ impl<F: CoordFloat + FromPrimitive> Bearing<F> for Haversine {
     ///
     /// ```
     /// # use approx::assert_relative_eq;
-    /// use geo::{HAVERSINE, Bearing};
+    /// use geo::{Haversine, Bearing};
     /// use geo::Point;
     ///
     /// let origin = Point::new(9.0, 10.0);
     /// let destination = Point::new(9.5, 10.1);
-    /// let bearing = HAVERSINE.bearing(origin, destination);
+    /// let bearing = Haversine.bearing(origin, destination);
     /// // A little north of east
     /// assert_relative_eq!(bearing, 78.47, epsilon = 1.0e-2);
     /// ```
@@ -169,7 +196,7 @@ impl<F: CoordFloat + FromPrimitive> Bearing<F> for Haversine {
     }
 }
 
-impl<F: CoordFloat + FromPrimitive> Destination<F> for Haversine {
+impl<F: CoordFloat + FromPrimitive> Destination<F> for HaversineMeasure {
     /// Returns a new point having travelled the `distance` along a [great circle]
     /// from the `origin` point with the given `bearing`.
     ///
@@ -184,20 +211,13 @@ impl<F: CoordFloat + FromPrimitive> Destination<F> for Haversine {
     ///
     /// ```
     /// # use approx::assert_relative_eq;
-    /// use geo::{HAVERSINE, Destination};
+    /// use geo::{Haversine, Destination};
     /// use geo::Point;
     ///
     /// let origin = Point::new(9.177789688110352, 48.776781529534965);
-    /// let destination = HAVERSINE.destination(origin, 45., 10000.);
+    /// let destination = Haversine.destination(origin, 45., 10000.);
     /// assert_relative_eq!(Point::new(9.274409949623532, 48.84033274015048), destination);
     /// ```
-    ///
-    /// # References
-    ///
-    /// *Note*: this implementation uses a mean earth radius of 6371.088 km, based on the [recommendation of
-    /// the IUGG](ftp://athena.fsv.cvut.cz/ZFG/grs80-Moritz.pdf)
-    ///
-    /// [great circle]: https://en.wikipedia.org/wiki/Great_circle
     fn destination(&self, origin: Point<F>, bearing: F, meters: F) -> Point<F> {
         let center_lng = origin.x().to_radians();
         let center_lat = origin.y().to_radians();
@@ -216,7 +236,7 @@ impl<F: CoordFloat + FromPrimitive> Destination<F> for Haversine {
     }
 }
 
-impl<F: CoordFloat + FromPrimitive> Distance<F, Point<F>, Point<F>> for Haversine {
+impl<F: CoordFloat + FromPrimitive> Distance<F, Point<F>, Point<F>> for HaversineMeasure {
     /// Determine the distance between two points using the [haversine formula].
     ///
     /// # Units
@@ -228,24 +248,19 @@ impl<F: CoordFloat + FromPrimitive> Distance<F, Point<F>, Point<F>> for Haversin
     ///
     /// ```
     /// # use approx::assert_relative_eq;
-    /// use geo::{HAVERSINE, Distance};
+    /// use geo::{Haversine, Distance};
     /// use geo::Point;
     ///
     /// let new_york_city = Point::new(-74.006f64, 40.7128f64);
     /// let london = Point::new(-0.1278f64, 51.5074f64);
     ///
-    /// let distance = HAVERSINE.distance(new_york_city, london);
+    /// let distance = Haversine.distance(new_york_city, london);
     ///
     /// assert_relative_eq!(
     ///     5_570_230., // meters
     ///     distance.round()
     /// );
     /// ```
-    ///
-    /// # References
-    ///
-    /// *Note*: this implementation uses a mean earth radius of 6371.088 km, based on the [recommendation of
-    /// the IUGG](ftp://athena.fsv.cvut.cz/ZFG/grs80-Moritz.pdf)
     ///
     /// [haversine formula]: https://en.wikipedia.org/wiki/Haversine_formula
     fn distance(&self, origin: Point<F>, destination: Point<F>) -> F {
@@ -264,23 +279,23 @@ impl<F: CoordFloat + FromPrimitive> Distance<F, Point<F>, Point<F>> for Haversin
 /// Interpolate Point(s) along a [great circle].
 ///
 /// [great circle]: https://en.wikipedia.org/wiki/Great_circle
-impl<F: CoordFloat + FromPrimitive> InterpolatePoint<F> for Haversine {
+impl<F: CoordFloat + FromPrimitive> InterpolatePoint<F> for HaversineMeasure {
     /// Returns a new Point along a [great circle] between two existing points.
     ///
     /// # Examples
     ///
     /// ```
     /// # use approx::assert_relative_eq;
-    /// use geo::{HAVERSINE, InterpolatePoint};
+    /// use geo::{Haversine, InterpolatePoint};
     /// use geo::Point;
     ///
     /// let p1 = Point::new(10.0, 20.0);
     /// let p2 = Point::new(125.0, 25.0);
     ///
-    /// let closer_to_p1 = HAVERSINE.point_at_distance_between(p1, p2, 100_000.0);
+    /// let closer_to_p1 = Haversine.point_at_distance_between(p1, p2, 100_000.0);
     /// assert_relative_eq!(closer_to_p1, Point::new(10.81, 20.49), epsilon = 1.0e-2);
     ///
-    /// let closer_to_p2 = HAVERSINE.point_at_distance_between(p1, p2, 10_000_000.0);
+    /// let closer_to_p2 = Haversine.point_at_distance_between(p1, p2, 10_000_000.0);
     /// assert_relative_eq!(closer_to_p2, Point::new(112.33, 30.57), epsilon = 1.0e-2);
     /// ```
     ///
@@ -301,19 +316,19 @@ impl<F: CoordFloat + FromPrimitive> InterpolatePoint<F> for Haversine {
     ///
     /// ```
     /// # use approx::assert_relative_eq;
-    /// use geo::{HAVERSINE, InterpolatePoint};
+    /// use geo::{Haversine, InterpolatePoint};
     /// use geo::Point;
     ///
     /// let p1 = Point::new(10.0, 20.0);
     /// let p2 = Point::new(125.0, 25.0);
     ///
-    /// let closer_to_p1 = HAVERSINE.point_at_ratio_between(p1, p2, 0.1);
+    /// let closer_to_p1 = Haversine.point_at_ratio_between(p1, p2, 0.1);
     /// assert_relative_eq!(closer_to_p1, Point::new(19.52, 25.27), epsilon = 1.0e-2);
     ///
-    /// let closer_to_p2 = HAVERSINE.point_at_ratio_between(p1, p2, 0.9);
+    /// let closer_to_p2 = Haversine.point_at_ratio_between(p1, p2, 0.9);
     /// assert_relative_eq!(closer_to_p2, Point::new(114.72, 29.65), epsilon = 1.0e-2);
     ///
-    /// let midpoint = HAVERSINE.point_at_ratio_between(p1, p2, 0.5);
+    /// let midpoint = Haversine.point_at_ratio_between(p1, p2, 0.5);
     /// assert_relative_eq!(midpoint, Point::new(65.87, 37.62), epsilon = 1.0e-2);
     /// ```
     ///
@@ -473,28 +488,28 @@ mod tests {
         fn north() {
             let origin = Point::new(0.0, 0.0);
             let destination = Point::new(0.0, 1.0);
-            assert_relative_eq!(0.0, HAVERSINE.bearing(origin, destination));
+            assert_relative_eq!(0.0, Haversine.bearing(origin, destination));
         }
 
         #[test]
         fn east() {
             let origin = Point::new(0.0, 0.0);
             let destination = Point::new(1.0, 0.0);
-            assert_relative_eq!(90.0, HAVERSINE.bearing(origin, destination));
+            assert_relative_eq!(90.0, Haversine.bearing(origin, destination));
         }
 
         #[test]
         fn south() {
             let origin = Point::new(0.0, 0.0);
             let destination = Point::new(0.0, -1.0);
-            assert_relative_eq!(180.0, HAVERSINE.bearing(origin, destination));
+            assert_relative_eq!(180.0, Haversine.bearing(origin, destination));
         }
 
         #[test]
         fn west() {
             let origin = Point::new(0.0, 0.0);
             let destination = Point::new(-1.0, 0.0);
-            assert_relative_eq!(270.0, HAVERSINE.bearing(origin, destination));
+            assert_relative_eq!(270.0, Haversine.bearing(origin, destination));
         }
     }
 
@@ -507,7 +522,7 @@ mod tests {
             let bearing = 0.0;
             assert_relative_eq!(
                 Point::new(0.0, 0.899320363724538),
-                HAVERSINE.destination(origin, bearing, 100_000.0)
+                Haversine.destination(origin, bearing, 100_000.0)
             );
         }
 
@@ -517,7 +532,7 @@ mod tests {
             let bearing = 90.0;
             assert_relative_eq!(
                 Point::new(0.8993203637245415, 5.506522912913066e-17),
-                HAVERSINE.destination(origin, bearing, 100_000.0)
+                Haversine.destination(origin, bearing, 100_000.0)
             );
         }
 
@@ -527,7 +542,7 @@ mod tests {
             let bearing = 180.0;
             assert_relative_eq!(
                 Point::new(0.0, -0.899320363724538),
-                HAVERSINE.destination(origin, bearing, 100_000.0)
+                Haversine.destination(origin, bearing, 100_000.0)
             );
         }
 
@@ -537,7 +552,7 @@ mod tests {
             let bearing = 270.0;
             assert_relative_eq!(
                 Point::new(-0.8993203637245415, -1.6519568738739197e-16),
-                HAVERSINE.destination(origin, bearing, 100_000.0)
+                Haversine.destination(origin, bearing, 100_000.0)
             );
         }
     }
@@ -550,7 +565,7 @@ mod tests {
             let new_york_city = Point::new(-74.006f64, 40.7128f64);
             let london = Point::new(-0.1278f64, 51.5074f64);
 
-            let distance = HAVERSINE.distance(new_york_city, london);
+            let distance = Haversine.distance(new_york_city, london);
 
             assert_relative_eq!(
                 5_570_230., // meters
@@ -565,7 +580,7 @@ mod tests {
         fn point_at_ratio_between_midpoint() {
             let start = Point::new(10.0, 20.0);
             let end = Point::new(125.0, 25.0);
-            let midpoint = HAVERSINE.point_at_ratio_between(start, end, 0.5);
+            let midpoint = Haversine.point_at_ratio_between(start, end, 0.5);
             assert_relative_eq!(midpoint, Point::new(65.87394172511485, 37.61809316888599));
         }
         #[test]
@@ -573,7 +588,7 @@ mod tests {
             let start = Point::new(10.0, 20.0);
             let end = Point::new(125.0, 25.0);
             let max_dist = 1000000.0; // meters
-            let route = HAVERSINE
+            let route = Haversine
                 .points_along_line(start, end, max_dist, true)
                 .collect::<Vec<_>>();
             assert_eq!(route.len(), 13);
@@ -586,7 +601,7 @@ mod tests {
             let start = Point::new(10.0, 20.0);
             let end = Point::new(125.0, 25.0);
             let max_dist = 1000000.0; // meters
-            let route = HAVERSINE
+            let route = Haversine
                 .points_along_line(start, end, max_dist, false)
                 .collect::<Vec<_>>();
             assert_eq!(route.len(), 11);